:sparkles: Add kinetic sdk package implementation (#1)
* Init and Setup implementation

* Add account creation

* Add MakeTransfer

* Implement SDK package

+++ Runtime/KinApiCodegen/KinApiClient/src/main/CsharpDotNet2/IO/Swagger/Model/AirdropStatsCounts.cs
@@ -17,21 +17,21 @@ namespace IO.Swagger.Model {
     /// </summary>
     [DataMember(Name="averageValue", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "averageValue")]
-    public decimal? AverageValue { get; set; }
+    public int? AverageValue { get; set; }
 
     /// <summary>
     /// Gets or Sets Total
     /// </summary>
     [DataMember(Name="total", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "total")]
-    public decimal? Total { get; set; }
+    public int? Total { get; set; }
 
     /// <summary>
     /// Gets or Sets TotalValue
     /// </summary>
     [DataMember(Name="totalValue", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "totalValue")]
-    public decimal? TotalValue { get; set; }
+    public int? TotalValue { get; set; }
 
 
     /// <summary>
