+++ Runtime/KinApiCodegen/KinApiClient/src/main/CsharpDotNet2/IO/Swagger/Model/AppConfigApp.cs
@@ -17,7 +17,7 @@ namespace IO.Swagger.Model {
     /// </summary>
     [DataMember(Name="index", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "index")]
-    public decimal? Index { get; set; }
+    public int? Index { get; set; }
 
     /// <summary>
     /// Gets or Sets Name
