+++ Runtime/KinApiCodegen/KinApiClient/src/main/CsharpDotNet2/IO/Swagger/Model/MinimumRentExemptionBalanceResponse.cs
@@ -17,7 +17,7 @@ namespace IO.Swagger.Model {
     /// </summary>
     [DataMember(Name="lamports", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "lamports")]
-    public decimal? Lamports { get; set; }
+    public float? Lamports { get; set; }
 
 
     /// <summary>
